implementacion de arquitectura
